requirements.txt
azure-identity==1.7.1
azure-core==1.24.1